requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
billiard==4.2.3
boto3==1.41.4
botocore==1.41.4
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jmespath==1.0.1
kombu==5.5.4
mysqlclient==2.2.7
openai==2.8.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==1.7.1
PyMuPDF==1.26.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
s3transfer==0.15.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
